requirements.txt
streamlit
pandas
folium
streamlit-folium
streamlit-image-comparison
Pillow
